Unglobalize short_circuit
In the Gittip test suite we want to vary the short_circuit behavior of
the Aspen Website algorithm. If that's done via an attribute on the
website object then we constrain ourselves to synchronous testing, which
is unwise. By storing a default_short_circuit on the Algorithm instance
that's configurable via the constructor, we can also expose a per-call
override. That's what we're doing here.